requirements.txt
# MCP Python SDK
mcp>=1.0.0

fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database storage
redis>=5.0.0
psycopg2-binary>=2.9.0

# HTTP client
httpx>=0.25.0
